+++ packages/adapter-base-ui/src/DataTable.branches.test.tsx
@@ -13,8 +13,9 @@ import { useDataTableShell } from "@adapttable/core/adapter";
 import { fireEvent, render, screen, waitFor } from "@testing-library/react";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
-import { FilterDrawer, LoadingState } from "./components/chrome";
+import { FilterDrawer } from "./components/FilterDrawer";
 import { FilterPopover } from "./components/FilterPopover";
+import { LoadingState } from "./components/TableSkeleton";
 import { DataTable } from "./DataTable";
 import type { ColumnDef } from "./index";
