+++ packages/adapter-base-ui/src/components/ExpandToggle.tsx
@@ -1,0 +1,31 @@
+/** The row-expansion chevron, shared by rows and cards. */
+import { type Direction, type TableLabels } from "@adapttable/core";
+import { ExpandChevron } from "@adapttable/core/adapter";
+
+import { IconButton } from "../ui";
+
+/** Chevron toggle for a row's detail panel. */
+export function ExpandToggle({
+  open,
+  dir,
+  labels,
+  onToggle,
+}: Readonly<{
+  open: boolean;
+  dir?: Direction;
+  labels: Pick<Required<TableLabels>, "expandRow" | "collapseRow">;
+  onToggle: () => void;
+}>) {
+  return (
+    <IconButton
+      size="1"
+      variant="ghost"
+      color="gray"
+      aria-expanded={open}
+      aria-label={open ? labels.collapseRow : labels.expandRow}
+      onClick={onToggle}
+    >
+      <ExpandChevron open={open} dir={dir} />
+    </IconButton>
+  );
+}
